Initial draft of MongoDB m to n

+++ deliveryService/src/main/java/edu/tum/ase/deliveryService/controller/BoxController.java
@@ -1,6 +1,10 @@
 package edu.tum.ase.deliveryService.controller;
 
+import edu.tum.ase.deliveryService.exceptions.BoxHasActiveDeliveriesException;
+import edu.tum.ase.deliveryService.exceptions.BoxHasDeliveredDeliveriesException;
 import edu.tum.ase.deliveryService.model.Box;
+import edu.tum.ase.deliveryService.model.Delivery;
+import edu.tum.ase.deliveryService.model.DeliveryStatus;
 import edu.tum.ase.deliveryService.service.BoxService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Required;
@@ -9,6 +13,7 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.Collection;
 import java.util.List;
 
 @RestController
@@ -42,6 +47,22 @@ public class BoxController {
 
     //##################################################################################################################
     // PUT mappings
+    @PutMapping("{id}")
+    @PreAuthorize("hasRole('DISPATCHER')")
+    public Box updateBox(@RequestBody Box newBox, @PathVariable String id) {
+        Box box = boxService.findById(id);
+
+        // Check for not yet delivered "deliveries"
+        Collection<Delivery> deliveries = box.getDeliveries();
+        for (Delivery delivery : deliveries) {
+            if (!delivery.getStatus().equals(DeliveryStatus.DELIVERED)){
+                throw new BoxHasDeliveredDeliveriesException();
+            }
+        }
+
+        newBox.setId(id);
+        return boxService.updateBox(newBox);
+    }
 
     //##################################################################################################################
     // DELETE mappings
@@ -50,9 +71,14 @@ public class BoxController {
     @PreAuthorize("hasRole('DISPATCHER')")
     public HttpStatus deleteBox(@PathVariable String id) {
         Box box = boxService.findById(id);
-        boxService.delete(box);
 
+        // Check for active deliveries
+        Collection<Delivery> deliveries = box.getDeliveries();
+        if (deliveries.size() > 0) {
+            throw new BoxHasActiveDeliveriesException();
+        } else {
+            boxService.delete(box);
+        }
         return HttpStatus.OK;
     }
-
 }
